Move CCT to Core.Analysis
Implements #4796
